add can not delete usergroup

+++ src/Controller/UsergroupController.php
@@ -97,6 +97,7 @@ class UsergroupController extends ZhyuController
         try {
             $model = $this->repository->find($id);
 
+            @$title = (string) $model;
             return parent::view(null, $model, ['title' => $title]);
         }catch (\Exception $e){
 
@@ -145,6 +146,9 @@ class UsergroupController extends ZhyuController
     public function destroy($id)
     {
         $this->authorize('admin.usergroups.destroy');
+
+        $can_not_delete_ids = exploade(',', env('ZHYU_ADMIN_GROUP_CAN_NOT_DELETE'));
+        abort_if(in_array($id, $can_not_delete_ids), 403);
 
         $this->repository->delete($id);
